Fix MEDIUM issues #55-#68 from code review
Execution (#55-#61): Log exit stop check exceptions, document abs()
slippage intent, guard correlation threshold >= 1.0, replace pytz with
zoneinfo, use public post_webhook API, derive repo_root from __file__,
make _normalize_constraints recursive.

Root (#62-#68): Use float64 for OHLC in cache_store, document backtest
complexity, pass bad_tickers explicitly in scan_engine, fix config
comment (3% not 1%), move alpaca-py to requirements.txt, fix sentinel
dotenv path and suppress deprecated execution.py warning.

+++ cache_store.py
@@ -46,12 +46,12 @@ def write_parquet(df: pd.DataFrame, path: str):
     if df is None or df.empty:
         return
 
-    # SOP: Downcast types to save memory/space
+    # SOP: Downcast types to save memory/space (float64 preserves precision for all price levels)
     df = df.copy()
     float_cols = ["Open", "High", "Low", "Close"]
     for col in float_cols:
         if col in df.columns:
-            df[col] = df[col].astype("float32")
+            df[col] = df[col].astype("float64")
     
     if "Volume" in df.columns:
         df["Volume"] = df["Volume"].astype("float64")
